agrego funcionalidad de roles y permisos en navbar, crud paciente y atencion

+++ resources/views/guardTeamUser/attention/show.blade.php
@@ -6,7 +6,10 @@
 		<h1 class="page-header">Atención al Paciente</h1>
 	</div>
     <div class="col-sm-12">
-        <a href="#" class="btn btn-primary float-right" data-toggle="modal" data-target="#create" data-backdrop="static" data-keyboard="false">Nueva Atención</a>
+        @can('attention_new')
+            @include('guardTeamUser.attention.create')
+            <a href="#" class="btn btn-primary float-right" data-toggle="modal" data-target="#create" data-backdrop="static" data-keyboard="false">Nueva Atención</a>
+        @endcan
         <table class="table table-hover table-striped mt-4">
             <thead>
                 <tr>
@@ -22,21 +25,26 @@
                     <td class="text-center">@{{ a.patient.surname }}</td>
                     <td class="text-center">@{{ a.diagnostic }}</td>
                     <td class="text-center">@{{ a.date }}</td>
-                    <td width="10px">
-                        <a href="#" class="btn btn-info btn-sm" v-on:click.prevent="detailsAttention(a)">Detalles</a>
-                    </td>
-                    <td width="10px">
-                        <a href="#" class="btn btn-warning btn-sm" v-on:click.prevent="editAttention(a)">Editar</a>
-                    </td>
-                    <td width="10px">
-                        <a href="#" class="btn btn-danger btn-sm" v-on:click.prevent="deleteAttention(a)">Eliminar</a>
-                    </td>
+                    @can('attention_show') 
+                        @include('guardTeamUser.attention.details')
+                        <td width="10px">
+                            <a href="#" class="btn btn-info btn-sm" v-on:click.prevent="detailsAttention(a)">Detalles</a>
+                        </td>
+                    @endcan
+                    @can('attention_update')
+                        @include('guardTeamUser.attention.edit')
+                        <td width="10px">
+                            <a href="#" class="btn btn-warning btn-sm" v-on:click.prevent="editAttention(a)">Editar</a>
+                        </td>
+                    @endcan
+                    @can('attention_destroy')
+                        <td width="10px">
+                            <a href="#" class="btn btn-danger btn-sm" v-on:click.prevent="deleteAttention(a)">Eliminar</a>
+                        </td>
+                    @endcan
                 </tr>
             </tbody>
         </table>
     </div>
-    @include('guardTeamUser.attention.create')
-    @include('guardTeamUser.attention.edit')
-    @include('guardTeamUser.attention.details')
 </div>
 @endsection
